Update CRUD follow format

+++ DataAccess/Entities/University.cs
@@ -11,7 +11,7 @@ namespace DataAccess.Entities
         public string Name { get; set; }
         public string ShortName { get; set; }
         public string Location { get; set; }
-        public int Ranking { get; set; }
+        public double Ranking { get; set; }
         public string? Avatar { get; set; }
     }
 }
